Speed Improvement
I sped up the algorithm 110x, instead of traversing through the
coordinates to determine if a coord is open / closed. I cloned the
original map and set all the data in the coord to 1 (open) so I can now
access the available coords like: available[row][col].

For the record I knew this was the correct way to do it in the first
place but I write sloppy code when I stare at it for too long. next
time, I should take a break and think about what I am doing.